[DependencyInjection] Move the CACHEDIR.TAG test to the component owning the feature
Writing CACHEDIR.TAG happens in KernelTrait, which lives in
DependencyInjection, but the test sat in HttpKernel. HttpKernel allows
symfony/dependency-injection ^8.1, so its low-deps job installs 8.1,
where the feature does not exist yet, and the test fails there.
